+++ _utils/_notebook_utils/z_rsync_utils.py
@@ -1,3 +1,0 @@
-"""Helpers for notebooks/Z_rsync.ipynb."""
-
-__all__ = []